workload/pyproject.toml
# Copyright 2026 Tony Meyer
# See LICENSE file for licensing details.

[project]
name = "mcp-server-workload"
version = "0.0.1"
requires-python = ">=3.10"
description = "MCP server workload for the mcp-server Juju charm"

dependencies = [
    "mcp[cli]",
    "PyJWT[crypto]",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
    "pytest",
    "coverage[toml]",
]

# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"
pythonpath = ["src"]

[tool.ty.environment]
extra-paths = ["src"]

# Linting tools configuration
[tool.ruff]
line-length = 99
lint.select = ["E", "W", "F", "C", "N", "D", "I001"]
lint.ignore = [
    "D105",
    "D107",
    "D203",
    "D204",
    "D213",
    "D215",
    "D400",
    "D404",
    "D406",
    "D407",
    "D408",
    "D409",
    "D413",
]
extend-exclude = ["__pycache__", "*.egg_info"]
lint.per-file-ignores = {"tests/*" = ["D100","D101","D102","D103","D104"]}

[tool.ruff.lint.mccabe]
max-complexity = 10
